meddy_gpt/requirements.txt
langchain_openai 
langchain_core
python-dotenv
langchain_community
pypdf
langchainhub
sentence_transformers
PyPDF2
langchain_pinecone
flask
flask-cors
